+++ DataFormats/interface/SecondaryVertexFeatures.h
@@ -13,7 +13,7 @@ class SecondaryVertexFeatures {
     float sv_mass;
     float sv_ntracks;
     float sv_chi2;
-    float sv_ndof;
+    int sv_ndof;
     float sv_dxy;
     float sv_dxysig;
     float sv_d3d;
@@ -25,4 +25,4 @@ class SecondaryVertexFeatures {
 
 }
 
-#endif 
+#endif
